requirements.txt
torch>=2.9.1
transformers>=4.57.1
numpy>=2.2.6,<2.3.0
datasets>=4.4.1
jinja2>=3.1.6
faiss-cpu>=1.12.0
sentence-transformers>=5.1.2
fastapi>=0.121.2
uvicorn[standard]>=0.38.0
pydantic>=2.12.4
pydantic-settings>=2.12.0
httpx[http2]>=0.28.1
msgspec>=0.19.0
orjson>=3.11.4
zstandard>=0.25.0
lz4>=4.4.5
prometheus-client>=0.23.1
opentelemetry-instrumentation-fastapi>=0.59b0
opentelemetry-instrumentation-httpx>=0.59b0
opentelemetry-exporter-otlp>=1.38.0
opentelemetry-sdk>=1.38.0
psutil>=7.1.3
scalene>=1.5.55
pyarrow>=22.0.0
safetensors>=0.6.2
diskcache>=5.6.3
tenacity>=9.1.2
requests>=2.32.5
pandas>=2.3.3
matplotlib>=3.10.7
seaborn>=0.13.2
pyyaml>=6.0.3
